[ci] move global configs to new configs.py (#57190)
this makes it accessible from all places
