+++ ci/ray_ci/docker_container.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from enum import Enum
 from typing import Dict, List
 
-from ci.ray_ci.builder_container import DEFAULT_ARCHITECTURE, DEFAULT_PYTHON_VERSION
+from ci.ray_ci.configs import DEFAULT_ARCHITECTURE, DEFAULT_PYTHON_VERSION
 from ci.ray_ci.linux_container import LinuxContainer
 
 PLATFORMS_RAY = [
