+++ ci/ray_ci/test_ray_docker_container.py
@@ -7,7 +7,7 @@ from unittest import mock
 import pytest
 from ray_release.configs.global_config import get_global_config
 
-from ci.ray_ci.builder_container import DEFAULT_PYTHON_VERSION
+from ci.ray_ci.configs import DEFAULT_PYTHON_VERSION
 from ci.ray_ci.container import _DOCKER_ECR_REPO
 from ci.ray_ci.docker_container import GPU_PLATFORM
 from ci.ray_ci.ray_docker_container import RayDockerContainer
